# day-2/book-management-api/openapi/openapi.yaml
openapi: 3.1.0

info:
  title: Book Management API
  version: 1.0.0
  description: >
    Contract-First REST API for managing books.
    This API demonstrates CRUD operations using OpenAPI 3.1.

jsonSchemaDialect: https://json-schema.org/draft/2020-12/schema

servers:
  - url: http://localhost:3000/api/v1
    description: Local Development Server

tags:
  - name: Books
    description: Book Management Operations

paths:
  /books:
    get:
      tags:
        - Books
      summary: Get all books
      operationId: getBooks
      description: Returns a list of all books.
      responses:
        "200":
          description: List of books retrieved successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookListResponse"

        "500":
          $ref: "#/components/responses/InternalServerError"

    post:
      tags:
        - Books
      summary: Create a new book
      operationId: createBook

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookRequest"

      responses:
        "201":
          description: Book created successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookResponse"

        "400":
          $ref: "#/components/responses/BadRequest"

        "500":
          $ref: "#/components/responses/InternalServerError"

  /books/{id}:

    parameters:
      - $ref: "#/components/parameters/BookId"

    get:
      tags:
        - Books
      summary: Get a book by ID
      operationId: getBookById

      responses:
        "200":
          description: Book found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookResponse"

        "404":
          $ref: "#/components/responses/NotFound"

        "500":
          $ref: "#/components/responses/InternalServerError"

    put:
      tags:
        - Books
      summary: Update an existing book
      operationId: updateBook

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBookRequest"

      responses:
        "200":
          description: Book updated successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookResponse"

        "400":
          $ref: "#/components/responses/BadRequest"

        "404":
          $ref: "#/components/responses/NotFound"

        "500":
          $ref: "#/components/responses/InternalServerError"

    delete:
      tags:
        - Books
      summary: Delete a book
      operationId: deleteBook

      responses:
        "204":
          description: Book deleted successfully.

        "404":
          $ref: "#/components/responses/NotFound"

        "500":
          $ref: "#/components/responses/InternalServerError"

components:

  parameters:

    BookId:
      name: id
      in: path
      required: true
      description: Unique Book Identifier
      schema:
        type: integer
        minimum: 1
        example: 101

  schemas:

    Book:
      type: object

      required:
        - id
        - title
        - author
        - isbn
        - genre
        - publishedYear
        - price
        - available
        - createdAt
        - updatedAt

      properties:

        id:
          type: integer
          example: 101

        title:
          type: string
          minLength: 2
          maxLength: 200
          example: Clean Code

        author:
          type: string
          example: Robert C. Martin

        isbn:
          type: string
          example: "9780132350884"

        genre:
          type: string
          example: Software Engineering

        publishedYear:
          type: integer
          minimum: 1900
          maximum: 2100
          example: 2008

        price:
          type: number
          format: double
          minimum: 0
          example: 699.99

        available:
          type: boolean
          example: true

        createdAt:
          type: string
          format: date-time
          example: "2026-07-02T10:00:00Z"

        updatedAt:
          type: string
          format: date-time
          example: "2026-07-02T10:00:00Z"

    CreateBookRequest:
      type: object

      required:
        - title
        - author
        - isbn
        - genre
        - publishedYear
        - price

      properties:

        title:
          type: string
          minLength: 2
          maxLength: 200

        author:
          type: string

        isbn:
          type: string

        genre:
          type: string

        publishedYear:
          type: integer
          minimum: 1900

        price:
          type: number
          minimum: 0

        available:
          type: boolean
          default: true

    UpdateBookRequest:
      type: object

      properties:

        title:
          type: string

        author:
          type: string

        isbn:
          type: string

        genre:
          type: string

        publishedYear:
          type: integer

        price:
          type: number

        available:
          type: boolean

    ApiResponse:
      type: object

      required:
        - success
        - message
        - timestamp

      properties:

        success:
          type: boolean
          example: true

        message:
          type: string
          example: Request processed successfully.

        timestamp:
          type: string
          format: date-time

    BookResponse:
      allOf:
        - $ref: "#/components/schemas/ApiResponse"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/Book"

    BookListResponse:
      allOf:
        - $ref: "#/components/schemas/ApiResponse"
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/Book"

    ErrorResponse:
      type: object

      required:
        - success
        - message
        - timestamp
        - error

      properties:

        success:
          type: boolean
          example: false

        message:
          type: string
          example: Validation failed.

        timestamp:
          type: string
          format: date-time

        error:
          type: object

          required:
            - status
            - code
            - details

          properties:

            status:
              type: integer
              example: 400

            code:
              type: string
              example: BAD_REQUEST

            details:
              type: string
              example: Title is required.

  responses:

    BadRequest:
      description: Invalid request.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

    NotFound:
      description: Book not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

    InternalServerError:
      description: Unexpected server error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"